// apps/core/src/payload-types.ts
export interface Tenant {
  id: string;
  name: string;
  /**
   * Used for domain-based tenant handling
   */
  domain?: string | null;
  /**
   * Used for url paths, example: /tenant-slug/page-slug
   */
  slug: string;
  /**
   * If checked, logging in is not required to read. Useful for building public pages.
   */
  allowPublicRead?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  tenant?: (string | null) | Tenant;
  /**
   * Search engines use alt text to understand the content of an image, as they cannot "see" images themselves.
   */
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    tablet?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    desktop?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    seo?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface Configuration {
  id: string;
  tenant?: (string | null) | Tenant;
  contact?: {
    schedule?:
      | {
          day?: string | null;
          time?: string | null;
          id?: string | null;
        }[]
      | null;
    phone?:
      | {
          type?: ('phone' | 'whatsapp' | 'email') | null;
          link?: string | null;
          label?: string | null;
          id?: string | null;
        }[]
      | null;
    address?: string | null;
    maps?: string | null;
  };
  social?: {
    facebook?: string | null;
    instagram?: string | null;
    youtube?: string | null;
    tiktok?: string | null;
    restaurantguru?: string | null;
    yelp?: string | null;
    tripadvisor?: string | null;
  };
  navigation?: {
    logo?: (string | null) | Media;
    sticky?: boolean | null;
    transparent?: boolean | null;
    heroMode?: boolean | null;
    borderBottom?: boolean | null;
    links?:
      | {
          link?: {
            label?: string | null;
            newTab?: boolean | null;
            type?: ('reference' | 'custom') | null;
            reference?: {
              relationTo: 'pages';
              value: string | Page;
            } | null;
            url?: string | null;
          };
          id?: string | null;
        }[]
      | null;
    actions?:
      | {
          link?: {
            label?: string | null;
            newTab?: boolean | null;
            type?: ('reference' | 'custom') | null;
            reference?: {
              relationTo: 'pages';
              value: string | Page;
            } | null;
            url?: string | null;
          };
          id?: string | null;
        }[]
      | null;
  };
  footer?: {
    minimal?: boolean | null;
  };
  theme?: {
    background?: string | null;
    fontColor?: string | null;
    surface?: string | null;
    onSurface?: string | null;
    primary?: string | null;
    onPrimary?: string | null;
    secondary?: string | null;
    onSecondary?: string | null;
    error?: string | null;
    onError?: string | null;
    border?: string | null;
  };
  i18n?: {
    es?: boolean | null;
    en?: boolean | null;
    de?: boolean | null;
    'es-ar'?: boolean | null;
  };
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}